# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ClienteResponse:
    properties:
      cpf:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      nome:
        type: string
      telefone:
        type: string
      updated_at:
        type: string
    type: object
  dto.CountResponse:
    properties:
      count:
        type: integer
    type: object
  dto.CreateClienteRequest:
    properties:
      cpf:
        type: string
      email:
        type: string
      nome:
        maxLength: 100
        minLength: 3
        type: string
      telefone:
        maxLength: 15
        minLength: 10
        type: string
    required:
    - cpf
    - email
    - nome
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.UpdateClienteRequest:
    properties:
      cpf:
        type: string
      email:
        type: string
      nome:
        maxLength: 100
        minLength: 3
        type: string
      telefone:
        maxLength: 15
        minLength: 10
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@exemplo.com
    name: API Support
  description: REST API for managing clientes (customers) with CRUD operations
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Cliente API
  version: "1.0"
paths:
  /clientes:
    get:
      description: Retrieve all clientes from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ClienteResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all clientes
      tags:
      - clientes
    post:
      consumes:
      - application/json
      description: Create a new cliente with the provided information
      parameters:
      - description: Cliente data
        in: body
        name: cliente
        required: true
        schema:
          $ref: '#/definitions/dto.CreateClienteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ClienteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new cliente
      tags:
      - clientes
  /clientes/{id}:
    delete:
      description: Delete a cliente by ID
      parameters:
      - description: Cliente ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete cliente
      tags:
      - clientes
    get:
      description: Retrieve a specific cliente by ID
      parameters:
      - description: Cliente ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ClienteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get cliente by ID
      tags:
      - clientes
    put:
      consumes:
      - application/json
      description: Update an existing cliente
      parameters:
      - description: Cliente ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cliente data
        in: body
        name: cliente
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateClienteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ClienteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update cliente
      tags:
      - clientes
  /clientes/count:
    get:
      description: Get the total number of clientes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CountResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Count clientes
      tags:
      - clientes
  /clientes/nome/{name}:
    get:
      description: Retrieve clientes matching the specified name (partial match)
      parameters:
      - description: Cliente name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ClienteResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get clientes by name
      tags:
      - clientes
...
swagger: "2.0"